Refactor to AbstractTemplateConstants which enables service specific constants
